loading框bug修复，增加fragment生命周期监听，增加切换fragment事务方法以及对应的基础哦底部控件，增加文件转换操作

+++ baselibrary/src/main/java/org/markensic/baselibrary/api/utils/DisPlayUtils.kt
@@ -32,7 +32,7 @@ object DisPlayUtils {
         return (physicsDm.density * size).toInt()
     }
 
-    fun px2dip(size: Int): Int {
-        return (size / physicsDm.density + 0.5f).toInt()
+    fun px2dip(size: Int): Float {
+        return size / physicsDm.density + 0.5f
     }
 }
